remove casts

+++ tests/units/solver/impl/test_wavefield_acoustic.cc
@@ -57,8 +57,8 @@ TEST_F(WavefieldAcousticTest, Constructor)
   // Verify data is correctly stored
   for (size_t i = 0; i < size1; ++i)
   {
-    EXPECT_FLOAT_EQ(wavefield.m_pnGlobalPrev(i), static_cast<float>(i));
-    EXPECT_FLOAT_EQ(wavefield.m_pnGlobalCurr(i), static_cast<float>(i * 2));
+    EXPECT_FLOAT_EQ(wavefield.m_pnGlobalPrev(i), i);
+    EXPECT_FLOAT_EQ(wavefield.m_pnGlobalCurr(i), i * 2);
   }
 }
 
@@ -130,8 +130,8 @@ TEST_F(WavefieldAcousticTest, CopyAssignmentSelfAssignment)
 
   for (size_t i = 0; i < size1; ++i)
   {
-    EXPECT_FLOAT_EQ(wavefield.m_pnGlobalPrev(i), static_cast<float>(i));
-    EXPECT_FLOAT_EQ(wavefield.m_pnGlobalCurr(i), static_cast<float>(i * 2));
+    EXPECT_FLOAT_EQ(wavefield.m_pnGlobalPrev(i), i);
+    EXPECT_FLOAT_EQ(wavefield.m_pnGlobalCurr(i), i * 2);
   }
 }
